feat : game 대기실 방 상세 정보 Get 구현

+++ src/main/java/com/movie/domain/game/dto/request/CreateGameReqDto.java
@@ -2,7 +2,7 @@ package com.movie.domain.game.dto.request;
 
 public record CreateGameReqDto(
         String title,
-        Integer maxPlayer
+        Long maxPlayer
 ) {
 
 }
